clean up manifest build function
